backend/requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
anyio==4.2.0
attrs==23.1.0
black==23.12.0
blessed==1.20.0
botocore==1.34.2
chalice==1.30.0
click==8.1.7
frozenlist==1.4.1
graphql-core==3.2.3
h11==0.14.0
idna==3.6
iniconfig==2.0.0
inquirer==2.10.1
jmespath==1.0.1
libcst==1.1.0
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.0.4
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.1.0
pluggy==1.3.0
Pygments==2.17.2
pytest==7.4.3
python-dateutil==2.8.2
python-editor==1.0.4
python-multipart==0.0.6
PyYAML==6.0.1
readchar==4.0.5
rich==13.7.0
six==1.16.0
sniffio==1.3.0
starlette==0.34.0
strawberry-graphql==0.216.1
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.9.0
urllib3==2.0.7
uvicorn==0.24.0.post1
wcwidth==0.2.12
yarl==1.9.4
